+++ packages/nyris-react-components/src/assets/types.ts
@@ -30,6 +30,7 @@ export const ICONS = {
   logout: 'logout',
   machine_view: 'machine_view',
   minus: 'minus',
+  mobile_camera: 'mobile_camera',
   next_arrow: 'next_arrow',
   part: 'part',
   plus_rounded: 'plus_rounded',